fix(flower): fix startUp bash script

+++ project/flower/quickstart-monai/start.sh
@@ -1,15 +1,16 @@
 #!/bin/bash
 
-docker compose down 
+if [ -f docker compose.yml ]; then
+    docker compose down 
+fi
 
 if [ "$#" -eq 0 ]; then
     echo "No number of nodes passed, defaulting to 2 clients."
-    python generate_compose.py
+    python generate_compose.py;
 else
-    python generate_compose.py $1
+    python generate_compose.py $1;
 fi
 
-docker compose up -d --build
-
-flwr run . local-deployment --stream
+docker compose up -d --build;
+flwr run . local-deployment --stream;
